backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.26.3 
pydantic
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
